Move validation into core spec
This commit moves validation concepts into the core Smithy spec while
leaving more specific linters in a linters guide.
