+++ docs/source/guides/index.rst
@@ -8,6 +8,7 @@ Smithy Guides
     :maxdepth: 2
 
     building-models/index
+    model-linters
     evolving-models
     style-guide
     converting-to-openapi
