+++ smithy-model/src/main/resources/META-INF/services/software.amazon.smithy.model.validation.ValidatorService
@@ -1,2 +1,2 @@
-software.amazon.smithy.model.validation.EmitEachSelectorValidator$Provider
-software.amazon.smithy.model.validation.EmitNoneSelectorValidator$Provider
+software.amazon.smithy.model.validation.linters.EmitEachSelectorValidator$Provider
+software.amazon.smithy.model.validation.linters.EmitNoneSelectorValidator$Provider
